requirements.txt
oandapyV20>=0.7.2
requests>=2.31    # transitive dep of oandapyV20 — not auto-installed
python-telegram-bot>=21
APScheduler>=3.10
python-dotenv>=1.0
pandas>=2.0
numpy>=1.24
matplotlib>=3.8
